+++ CHANGELOG.md
@@ -15,6 +15,9 @@
   deprecated
   ([PR #766](https://github.com/jasmin-lang/jasmin/pull/766)).
 
+- Register allocation can print liveness information (enable with `-pliveness`)
+  ([PR #749](https://github.com/jasmin-lang/jasmin/pull/749)).
+
 ## Bug fixes
 
 - The compiler no longer throws an exception when a required file does not exist